+++ cmd/disk.go
@@ -18,8 +18,7 @@ package cmd
 import (
 	"fmt"
 
-	"github.com/obviyus/goids/getinfo"
-
+	"github.com/obviyus/goids/fetch"
 	"github.com/spf13/cobra"
 )
 
@@ -29,7 +28,7 @@ var diskCmd = &cobra.Command{
 	Short: "List basic information about disk",
 	Long:  `List Total, Used and Free space on all disks.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		err := getinfo.DiskInfo()
+		err := fetch.DiskInfo()
 		if err != nil {
 			fmt.Println("Calling DiskInfo failed: ", err)
 		}
