+++ fetch/disk_test.go
@@ -1,15 +1,19 @@
-package tests
+package fetch
 
 import (
 	"testing"
-
-	"github.com/obviyus/goids/getinfo"
 )
 
 func TestDiskInfo(t *testing.T) {
-	got := getinfo.DiskInfo()
+	got := DiskInfo()
 
 	if got != nil {
 		t.Error("DiskInfo failed to run: ", got)
+	}
+}
+
+func BenchmarkDiskInfo(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		DiskInfo()
 	}
 }
